+++ src/main.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import BackgroundShader from './components/BackgroundShader'
 import { checkHMDBrowser } from './utils'
-import { ReceiverScreen } from './screens/Receiver'
-import { SenderScreen } from './screens/Sender'
+import { ReceiverScreen } from './screens/ReceiverScreen'
+import { SenderScreen } from './screens/SenderScreen'
 import { ToastDispatcherProvider } from './providers/ToastDispatcher'
 import styles from './styles.module.css'
