--- src/screens/Receiver/index.tsx
@@ -1,7 +1,0 @@
-import { CameraView } from './components/CameraView'
-
-export const ReceiverScreen = () => {
-  return (
-    <CameraView />
-  )
-}
