+++ src/screens/ReceiverScreen/components/CameraCard/styles.module.css
