+++ src/screens/ReceiverScreen/components/HistoryCard/styles.module.css
