+++ src/screens/ReceiverScreen/index.tsx
@@ -1,0 +1,7 @@
+import { CameraCard } from './components/CameraCard'
+
+export const ReceiverScreen = () => {
+  return (
+    <CameraCard />
+  )
+}
