+++ src/screens/SenderScreen/components/QRView/index.tsx
