+++ src/screens/SenderScreen/components/QRView/styles.module.css
